[fix]: fix compile bugs

+++ src/qfvm_clifford/clifford_simulator.h
@@ -119,7 +119,7 @@ template <size_t word_size> struct circuit_simulator {
     } else if (COLLAPSING_GATE == pair.first) {
       auto record = unpack_vector<1>(pair.second, sim_tableau, ci.targets, rng);
       if (record.has_value())
-        sim_record.record(ci.targets[0], ci.args[0], record.value());
+        sim_record.record(ci.targets[0], static_cast<size_t>(ci.args[0]), record.value());
     } else if (ERROR_QUBIT_GATE == pair.first) {
       std::bernoulli_distribution d(ci.args[0]);
       if (d(rng))
